Correciones del coloquio n°2

+++ src/dao/DatosPersonalesDAO.java
@@ -6,9 +6,9 @@ import modelo.DatosPersonales;
 
 public interface DatosPersonalesDAO {
 
-    public void insertar(Connection cx, String nombre, String apellido, int dni);
+    public void insertar(Connection cx, String nombre, String apellido, long dni);
 
     public List<DatosPersonales> listarTodos(Connection cx);
     
-    public DatosPersonales buscarPorDNI(Connection cx, int dni);
+    public DatosPersonales buscarPorDNI(Connection cx, long dni);
 }
